fixes from PR
