+++ .secrets.baseline
@@ -345,28 +345,28 @@
         "filename": "core_backend/tests/api/conftest.py",
         "hashed_secret": "9fb7fe1217aed442b04c0f5e43b5d5a7d3287097",
         "is_verified": false,
-        "line_number": 42
+        "line_number": 45
       },
       {
         "type": "Secret Keyword",
         "filename": "core_backend/tests/api/conftest.py",
         "hashed_secret": "767ef7376d44bb6e52b390ddcd12c1cb1b3902a4",
         "is_verified": false,
-        "line_number": 43
+        "line_number": 46
       },
       {
         "type": "Secret Keyword",
         "filename": "core_backend/tests/api/conftest.py",
         "hashed_secret": "70240b5d0947cc97447de496284791c12b2e678a",
         "is_verified": false,
-        "line_number": 48
+        "line_number": 51
       },
       {
         "type": "Secret Keyword",
         "filename": "core_backend/tests/api/conftest.py",
         "hashed_secret": "80fea3e25cb7e28550d13af9dfda7a9bd08c1a78",
         "is_verified": false,
-        "line_number": 49
+        "line_number": 52
       }
     ],
     "core_backend/tests/api/test.env": [
@@ -609,5 +609,5 @@
       }
     ]
   },
-  "generated_at": "2024-07-31T15:59:15Z"
+  "generated_at": "2024-08-01T13:19:44Z"
 }
